count: one window and one sparse boolean
BooleanWindow<Lead, Others, Optional, Excludes, Table> and
BooleanSparse<Lead, Probes, Excludes, Table> replace Window,
WindowDisjunction, SparseConjunction, SparseExclusion and the fill-side
disjunction that exclusion used as a lead; the dispatch moves out of make.cpp
into make_boolean.cpp, the makers into one file per strategy. A conjunction
under excludes probes its tail inline instead of through an erased lead node,
and the exclude check stays branchless.
