+++ libs/iresearch/include/iresearch/search/count/boolean_window.hpp
@@ -22,58 +22,81 @@
 
 #include <algorithm>
 #include <cstdint>
-#include <span>
+#include <tuple>
 #include <type_traits>
 #include <utility>
-#include <vector>
 
 #include "basics/empty.hpp"
-#include "iresearch/search/common/plan.hpp"
-#include "iresearch/search/common/posting_fill.hpp"
-#include "iresearch/search/common/resolve.hpp"
 #include "iresearch/search/common/table_filter.hpp"
 #include "iresearch/search/common/window.hpp"
 #include "iresearch/search/count/root.hpp"
-#include "iresearch/search/fill/leaves.hpp"
-#include "iresearch/search/states/term_state.hpp"
 #include "iresearch/utils/type_limits.hpp"
 
 namespace irs::count {
 
-template<typename Lead, typename Others, typename Excludes, typename Table>
-class Window : public Root {
+template<typename Lead, typename Others, typename Optional, typename Excludes,
+         typename Table>
+class BooleanWindow : public Root {
  public:
+  static constexpr bool kLead = !std::is_same_v<Lead, utils::Empty>;
   static constexpr bool kOthers = !std::is_same_v<Others, utils::Empty>;
+  static constexpr bool kOptional = !std::is_same_v<Optional, utils::Empty>;
   static constexpr bool kExcludes = !std::is_same_v<Excludes, utils::Empty>;
-  template<typename LeadArgs, typename OthersArgs, typename ExcludesArgs>
-  Window(Table table, std::piecewise_construct_t, LeadArgs&& lead,
-         OthersArgs&& others, ExcludesArgs&& excludes)
+  static_assert(kLead != kOptional);
+  static_assert(kLead || !kOthers);
+
+  template<typename LeadArgs, typename OthersArgs, typename OptionalArgs,
+           typename ExcludesArgs>
+  BooleanWindow(Table table, std::piecewise_construct_t, LeadArgs&& lead,
+                OthersArgs&& others, OptionalArgs&& optional,
+                ExcludesArgs&& excludes)
     : _lead{std::make_from_tuple<Lead>(std::forward<LeadArgs>(lead))},
       _others{std::make_from_tuple<Others>(std::forward<OthersArgs>(others))},
+      _optional{
+        std::make_from_tuple<Optional>(std::forward<OptionalArgs>(optional))},
       _excludes{
         std::make_from_tuple<Excludes>(std::forward<ExcludesArgs>(excludes))},
       _table{table} {}
+
+  BooleanWindow(BooleanWindow&&) = delete;
+  BooleanWindow& operator=(BooleanWindow&&) = delete;
 
   uint64_t Run() final {
     uint64_t total = 0;
     doc_id_t min = doc_limits::min();
 
     for (;;) {
+      if constexpr (kOptional) {
+        if (_optional.Exhausted()) {
+          return total;
+        }
+      }
       if (!_table.Skip(min)) {
         return total;
       }
       SDB_ASSERT(min <= doc_limits::eof() - search::kWindowDocs);
       const doc_id_t max = min + search::kWindowDocs;
 
-      auto next = _lead.FillOr(min, max, _mask.data());
-      if constexpr (kOthers) {
-        next = std::max(next, _others.Restrict(min, max, _mask.data()));
+      doc_id_t next;
+      uint64_t* words;
+      if constexpr (kLead) {
+        words = _mask.data();
+        next = _lead.FillOr(min, max, words);
+        if constexpr (kOthers) {
+          next = std::max(next, _others.Restrict(min, max, words));
+        }
+      } else {
+        next = _optional.Fill(min, max);
+        words = _optional.Words();
       }
       if constexpr (kExcludes) {
-        _excludes.Remove(min, max, _mask.data());
+        _excludes.Remove(min, max, words);
       }
 
-      total += _table.CountAndClear(min, _mask.data(), search::kWindowWords);
+      total += _table.CountAndClear(min, words, search::kWindowWords);
+      if constexpr (kOptional) {
+        _optional.Reset();
+      }
 
       if (doc_limits::eof(next)) {
         return total;
@@ -83,39 +106,12 @@ class Window : public Root {
   }
 
  private:
-  search::Scratch _mask{};
-  Lead _lead;
+  [[no_unique_address]] utils::Need<kLead, search::Scratch> _mask{};
+  [[no_unique_address]] Lead _lead;
   [[no_unique_address]] Others _others;
+  [[no_unique_address]] Optional _optional;
   [[no_unique_address]] Excludes _excludes;
   [[no_unique_address]] search::Narrowing<Table> _table;
 };
-
-template<typename Excludes, typename Term, typename ExcludesArgs>
-Root::ptr MakeWindowOfTerms(std::span<const Term> terms,
-                            const TermReader* field, const IndexInput& doc,
-                            ExcludesArgs&& excludes, const Context& ctx) {
-  SDB_ASSERT(terms.size() >= 2);
-  return search::ResolveInput(doc, [&]<typename Input> -> Root::ptr {
-    using Leaf = search::PostingFill<Input>;
-    using Others = fill::AndLeaves<Leaf>;
-    const auto& own = search::FieldOf(terms.front(), field);
-    const auto& front = search::CookieOf(terms.front());
-    return MakeShape<Window, Leaf, Others, Excludes>(
-      ctx, std::piecewise_construct,
-      std::forward_as_tuple(front, doc,
-                            front.docs_count != 1 && search::BoundsOf(own),
-                            front.docs_count != 1 && search::FreqOf(own)),
-      std::forward_as_tuple(
-        terms.size() - 1,
-        [&](Leaf& leaf, size_t i) {
-          const auto& other = search::FieldOf(terms[i + 1], field);
-          const auto& meta = search::CookieOf(terms[i + 1]);
-          leaf.Prepare(meta, doc,
-                       meta.docs_count != 1 && search::BoundsOf(other),
-                       meta.docs_count != 1 && search::FreqOf(other));
-        }),
-      std::forward<ExcludesArgs>(excludes));
-  });
-}
 
 }  // namespace irs::count
