+++ libs/iresearch/include/iresearch/search/count/make_masked.cpp
@@ -20,9 +20,10 @@
 
 #include <utility>
 
+#include "basics/empty.hpp"
 #include "iresearch/index/index_reader.hpp"
+#include "iresearch/search/count/boolean_sparse.hpp"
 #include "iresearch/search/count/plan.hpp"
-#include "iresearch/search/count/sparse_exclusion.hpp"
 #include "iresearch/search/filter.hpp"
 #include "iresearch/search/lead/impl.hpp"
 #include "iresearch/search/probe/mask_docs.hpp"
@@ -36,9 +37,9 @@ Root::ptr MakeMasked(const QueryBuilder& query, const Context& ctx) {
   if (!node) {
     return {};
   }
-  return MakeShape<SparseExclusion, lead::Erased, probe::MaskDocs>(
+  return MakeShape<BooleanSparse, lead::Erased, utils::Empty, probe::MaskDocs>(
     ctx, std::piecewise_construct, std::forward_as_tuple(std::move(node)),
-    std::forward_as_tuple(*docs_mask));
+    std::forward_as_tuple(), std::forward_as_tuple(*docs_mask));
 }
 
 }  // namespace irs::count
