+++ libs/iresearch/include/iresearch/search/count/make_subtract_conjunction.cpp
@@ -24,7 +24,7 @@
 
 #include "iresearch/index/index_reader.hpp"
 #include "iresearch/search/common/window_of.hpp"
-#include "iresearch/search/count/plan.hpp"
+#include "iresearch/search/count/make_boolean.hpp"
 #include "iresearch/search/count/subtract.hpp"
 
 namespace irs::count {
