+++ libs/iresearch/include/iresearch/search/count/make_subtract_disjunction.cpp
@@ -23,7 +23,7 @@
 #include <utility>
 
 #include "iresearch/index/index_reader.hpp"
-#include "iresearch/search/count/plan.hpp"
+#include "iresearch/search/count/make_boolean.hpp"
 #include "iresearch/search/count/subtract.hpp"
 
 namespace irs::count {
